Fix plugin context dependency on configured order
Previously plugins were loaded in the order
they appeared in the configuration file and at
the same time plug in context references were
resolved and loaded.

This led to the error that when a plug in required
a context from a not yet loaded plugin the loading
sequence failed.

Now plug in contexts are loaded lazily.